Create 24 random temps
This was a code challenge where I was to create a list of 24 random ints in a range. Then sort the list in ascending order. Make index 4 = 32 if not already 32. Slice the first 6 numbers in the list to create a new list. An finally print the new list in descending order.